Updated TdSharedConent.hpp, fixed fov and new effect
The old TdSharedContent.hpp only had the pistol and not the other ones. I manually added them without adding the functions.

With the updated file, I could properly fix the fov effect, it now doesn't mess up when using the sniper rifle.

Added new effect that display the game's variables on the screen. More is planned to be added and to support different resolutions